A Bash shell script testaverager.sh that will determine if a person has passed a class
It should take in as a command-line parameter a name (first and
last name) as well as a series of numbers. This list could be of any length. Your script should
start by calculating the average of the scores. If the result is below 70, it should output
“Sorry [name] but you will have to retake the class!”. If the result is 70 or greater,
congratulation the student with some message about succeeding at the class.